refactor(agent-ui): extract WorktreeDetailContent from WorktreeDetailPopover
